+++ pkgs/test/CHANGELOG.md
@@ -45,6 +45,7 @@
 * Increase SDK constraint to ^3.5.0.
 * Support running Node.js tests compiled with dart2wasm.
 * Allow `firefox` or `firefox-bin` executable name on macOS.
+* Use a DevTools URL instead of a defunct observatory URL.
 
 ## 1.25.8